requirements.txt
kubernetes
google-auth
google-cloud-container